+++ src/main/java/com/cn/hotelDemo/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.cn.hotelDemo.exception;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(Long id) {
+        super("User not found with ID: " + id);
+    }
+}